address comments and refactor query file structure
