+++ app/react/kubernetes/namespaces/types.ts
@@ -4,14 +4,3 @@ export interface Namespaces {
     IsSystem: boolean;
   };
 }
-
-export interface SelfSubjectAccessReviewResponse {
-  status: {
-    allowed: boolean;
-  };
-  spec: {
-    resourceAttributes: {
-      namespace: string;
-    };
-  };
-}
